requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.22.1
mock==3.0.5
